server/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
joblib
google-generativeai
pypdf
